Simplify code around PerformanceReporter
